refactor: move overdue contract test cases to calculate_test.go
